rename to _examples and create submodules for each example
