--- _example/example.go
@@ -1,26 +1,0 @@
-package example
-
-type Status int
-
-var (
-	StatusUnknown            Status = 0
-	StatusStarted            Status = 1
-	StatusReadTheDocs        Status = 2
-	StatusFollowedTheExample Status = 3
-	StatusCreatedAFunExample Status = 4
-)
-
-func (s Status) String() string {
-	switch s {
-	case StatusStarted:
-		return "Started"
-	case StatusReadTheDocs:
-		return "Read the docs"
-	case StatusFollowedTheExample:
-		return "Followed the example"
-	case StatusCreatedAFunExample:
-		return "Created a fun example"
-	default:
-		return "Unknown"
-	}
-}
