+++ _examples/callback/callback_test.go
@@ -1,25 +1,22 @@
-package callbacks_test
+package callback_test
 
 import (
 	"context"
 	"testing"
 
 	"github.com/luno/jettison/jtest"
-
 	"github.com/luno/workflow"
 	"github.com/luno/workflow/adapters/memrecordstore"
 	"github.com/luno/workflow/adapters/memrolescheduler"
 	"github.com/luno/workflow/adapters/memstreamer"
-	"github.com/luno/workflow/adapters/memtimeoutstore"
-	"github.com/luno/workflow/example"
-	"github.com/luno/workflow/example/callbacks"
+
+	"github.com/luno/workflow/_examples/callback"
 )
 
 func TestCallbackWorkflow(t *testing.T) {
-	wf := callbacks.ExampleWorkflow(callbacks.Deps{
+	wf := callback.ExampleWorkflow(callback.Deps{
 		EventStreamer: memstreamer.New(),
 		RecordStore:   memrecordstore.New(),
-		TimeoutStore:  memtimeoutstore.New(),
 		RoleScheduler: memrolescheduler.New(),
 	})
 	t.Cleanup(wf.Stop)
@@ -28,14 +25,14 @@ func TestCallbackWorkflow(t *testing.T) {
 	wf.Run(ctx)
 
 	foreignID := "andrew"
-	runID, err := wf.Trigger(ctx, foreignID, example.StatusStarted)
+	runID, err := wf.Trigger(ctx, foreignID, callback.StatusStarted)
 	jtest.RequireNil(t, err)
 
-	workflow.TriggerCallbackOn(t, wf, foreignID, runID, example.StatusStarted, callbacks.EmailConfirmationResponse{
+	workflow.TriggerCallbackOn(t, wf, foreignID, runID, callback.StatusStarted, callback.EmailConfirmationResponse{
 		Confirmed: true,
 	})
 
-	workflow.Require(t, wf, foreignID, example.StatusFollowedTheExample, callbacks.Example{
+	workflow.Require(t, wf, foreignID, callback.StatusFollowedTheExample, callback.Example{
 		EmailConfirmed: true,
 	})
 }
