+++ _examples/schedule/go.mod
@@ -1,0 +1,33 @@
+module github.com/luno/workflow/_examples/schedule
+
+go 1.22.3
+
+replace github.com/luno/workflow => ../..
+
+require (
+	github.com/luno/jettison v0.0.0-20240722160230-b42bd507a5f6
+	github.com/luno/workflow v0.1.2
+	github.com/stretchr/testify v1.9.0
+	k8s.io/utils v0.0.0-20240711033017-18e509b52bc8
+)
+
+require (
+	github.com/beorn7/perks v1.0.1 // indirect
+	github.com/cespare/xxhash/v2 v2.2.0 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/go-stack/stack v1.8.1 // indirect
+	github.com/golang/protobuf v1.5.4 // indirect
+	github.com/google/uuid v1.3.0 // indirect
+	github.com/matttproud/golang_protobuf_extensions v1.0.4 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/prometheus/client_golang v1.15.0 // indirect
+	github.com/prometheus/client_model v0.3.0 // indirect
+	github.com/prometheus/common v0.42.0 // indirect
+	github.com/prometheus/procfs v0.9.0 // indirect
+	github.com/robfig/cron/v3 v3.0.0 // indirect
+	golang.org/x/sys v0.19.0 // indirect
+	golang.org/x/xerrors v0.0.0-20231012003039-104605ab7028 // indirect
+	google.golang.org/protobuf v1.34.2 // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
