+++ _examples/timeout/go.mod
@@ -1,14 +1,13 @@
-module github.com/luno/workflow/example
+module github.com/luno/workflow/_examples/timeout
 
 go 1.22.3
 
-replace github.com/luno/workflow => ..
+replace github.com/luno/workflow => ../..
 
 require (
-	github.com/luno/jettison v0.0.0-20230912135954-09d6084f5df9
-	github.com/luno/workflow v0.0.0-20240716150028-e4650a2adf6a
-	github.com/stretchr/testify v1.8.3
-	k8s.io/utils v0.0.0-20230505201702-9f6742963106
+	github.com/luno/jettison v0.0.0-20240722160230-b42bd507a5f6
+	github.com/luno/workflow v0.1.2
+	k8s.io/utils v0.0.0-20240711033017-18e509b52bc8
 )
 
 require (
@@ -16,7 +15,7 @@ require (
 	github.com/cespare/xxhash/v2 v2.2.0 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/go-stack/stack v1.8.1 // indirect
-	github.com/golang/protobuf v1.5.3 // indirect
+	github.com/golang/protobuf v1.5.4 // indirect
 	github.com/google/uuid v1.3.0 // indirect
 	github.com/matttproud/golang_protobuf_extensions v1.0.4 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
@@ -25,8 +24,9 @@ require (
 	github.com/prometheus/common v0.42.0 // indirect
 	github.com/prometheus/procfs v0.9.0 // indirect
 	github.com/robfig/cron/v3 v3.0.0 // indirect
-	golang.org/x/sys v0.9.0 // indirect
-	golang.org/x/xerrors v0.0.0-20220907171357-04be3eba64a2 // indirect
+	github.com/stretchr/testify v1.9.0 // indirect
+	golang.org/x/sys v0.19.0 // indirect
+	golang.org/x/xerrors v0.0.0-20231012003039-104605ab7028 // indirect
 	google.golang.org/protobuf v1.34.2 // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
