+++ _examples/timeout/timeout_test.go
@@ -1,4 +1,4 @@
-package timeouts_test
+package timeout_test
 
 import (
 	"context"
@@ -6,21 +6,20 @@ import (
 	"time"
 
 	"github.com/luno/jettison/jtest"
-	clocktesting "k8s.io/utils/clock/testing"
-
 	"github.com/luno/workflow"
 	"github.com/luno/workflow/adapters/memrecordstore"
 	"github.com/luno/workflow/adapters/memrolescheduler"
 	"github.com/luno/workflow/adapters/memstreamer"
 	"github.com/luno/workflow/adapters/memtimeoutstore"
-	"github.com/luno/workflow/example"
-	"github.com/luno/workflow/example/timeouts"
+	clocktesting "k8s.io/utils/clock/testing"
+
+	"github.com/luno/workflow/_examples/timeout"
 )
 
 func TestTimeoutWorkflow(t *testing.T) {
 	now := time.Now().UTC()
 	clock := clocktesting.NewFakeClock(now)
-	wf := timeouts.ExampleWorkflow(timeouts.Deps{
+	wf := timeout.ExampleWorkflow(timeout.Deps{
 		EventStreamer: memstreamer.New(),
 		RecordStore:   memrecordstore.New(),
 		TimeoutStore:  memtimeoutstore.New(),
@@ -33,14 +32,14 @@ func TestTimeoutWorkflow(t *testing.T) {
 	wf.Run(ctx)
 
 	foreignID := "andrew"
-	runID, err := wf.Trigger(ctx, foreignID, example.StatusStarted)
+	runID, err := wf.Trigger(ctx, foreignID, timeout.StatusStarted)
 	jtest.RequireNil(t, err)
 
-	workflow.AwaitTimeoutInsert(t, wf, foreignID, runID, example.StatusStarted)
+	workflow.AwaitTimeoutInsert(t, wf, foreignID, runID, timeout.StatusStarted)
 
 	clock.Step(time.Hour)
 
-	workflow.Require(t, wf, foreignID, example.StatusFollowedTheExample, timeouts.Example{
+	workflow.Require(t, wf, foreignID, timeout.StatusFollowedTheExample, timeout.Example{
 		Now: clock.Now(),
 	})
 }
